refactor: DeleteHistory constructor
