+++ src/test/java/qna/service/QnaServiceTest.java
@@ -1,14 +1,12 @@
 package qna.service;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import qna.CannotDeleteException;
 import qna.domain.*;
 
@@ -19,11 +17,9 @@ import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-@DataJpaTest
 @ExtendWith(MockitoExtension.class)
 class QnaServiceTest {
     @Mock
@@ -35,9 +31,6 @@ class QnaServiceTest {
     @Mock
     private DeleteHistoryService deleteHistoryService;
 
-    @Autowired
-    private UserRepository userRepository;
-
     @InjectMocks
     private QnaService qnaService;
 
@@ -47,26 +40,15 @@ class QnaServiceTest {
     @BeforeEach
     public void setUp() throws Exception {
         question = new Question(1L, "title1", "contents1").writeBy(UserTest.JAVAJIGI);
+        System.out.println("question : " + question.toString());
         answer = new Answer(1L, UserTest.JAVAJIGI, question, "Answers Contents1");
         question.addAnswer(answer);
     }
 
     @Test
-    @DisplayName("질문 등록")
-    void addQuestion() {
-        Question newQuestion = new Question("First Question", "This is first Question.");
-        User user = userRepository.save(new User("inmookjeong", "password", "inmookjeong", "jeonginmook@gmail.com"));
-        newQuestion.writeBy(user);
-        assertAll(
-                () -> assertThat(newQuestion).isNotNull(),
-                () -> assertThat(newQuestion.getTitle()).isEqualTo("First Question"),
-                () -> assertThat(newQuestion.getContents()).isEqualTo("This is first Question."),
-                () -> assertThat(newQuestion.getWriterId()).isEqualTo(user.getId())
-        );
-    }
-
-    @Test
     public void delete_성공() throws Exception {
+        System.out.println("questionId : " + question.getId());
+        System.out.println("qustino Ans : " + answerRepository.findByQuestionIdAndDeletedFalse(question.getId()).toString());
         when(questionRepository.findByIdAndDeletedFalse(question.getId())).thenReturn(Optional.of(question));
         when(answerRepository.findByQuestionIdAndDeletedFalse(question.getId())).thenReturn(Arrays.asList(answer));
 
@@ -111,8 +93,8 @@ class QnaServiceTest {
 
     private void verifyDeleteHistories() {
         List<DeleteHistory> deleteHistories = Arrays.asList(
-                new DeleteHistory(ContentType.QUESTION, question.getId(), question.getWriterId(), LocalDateTime.now()),
-                new DeleteHistory(ContentType.ANSWER, answer.getId(), answer.getWriterId(), LocalDateTime.now())
+                new DeleteHistory(ContentType.QUESTION, question.getId(), question.getWriter(), LocalDateTime.now()),
+                new DeleteHistory(ContentType.ANSWER, answer.getId(), answer.getWriter(), LocalDateTime.now())
         );
         verify(deleteHistoryService).saveAll(deleteHistories);
     }
